requirements.txt
colorama
peewee
telethon
requests
PyYAML
tiktoken
openai
sqlalchemy